Initial commit for crocodoc-ruby
This is the initial commit for a ruby client lib that interacts with v2 of the
Crocodoc API (crocodoc.com). It is a very thin wrapper around the api, cleaning
up a few rough edges, and raising ruby errors when it receives non-success
responses.

There is also a fake server implementation, which tries to mirror the Crocodoc
endpoints in the way it accepts parameters and returns responses.  It is
currently used for testing.